+++ commands/Tools/Step 3 - Modify schematisation/control_structures.py
@@ -13,8 +13,6 @@ from PyQt4.QtGui import QWidget
 
 from ThreeDiToolbox.commands.base.custom_command import CustomCommandBase
 from ThreeDiToolbox.threedi_schema_edits.controlled_structures import \
-    MEASURE_VARIABLE_WATERLEVEL
-from ThreeDiToolbox.threedi_schema_edits.controlled_structures import \
     RULE_OPERATOR_BOTTOM_UP
 from ThreeDiToolbox.threedi_schema_edits.controlled_structures import \
     RULE_OPERATOR_TOP_DOWN
@@ -181,6 +179,11 @@ class CustomCommand(CustomCommandBase):
         self.dockwidget_controlled_structures\
             .pushbutton_input_measuring_group_view.clicked.connect(
                 self.view_measuring_group)
+        self.dockwidget_controlled_structures\
+            .pushbutton_input_measuring_group_clear.clicked.connect(
+                self.remove_all_measuring_group_tabs)
+        self.dockwidget_controlled_structures.tab_measuring_group_view_2\
+            .tabCloseRequested.connect(self.remove_measuring_group_tab)
 
     def create_new_measuring_point(self):
         """Create a new measuring point."""
@@ -359,3 +362,14 @@ class CustomCommand(CustomCommandBase):
         self.dockwidget_controlled_structures\
             .tab_measuring_group_view_2.insertTab(0, tab, "Group: {}".format(
                 str(measuring_group_id)))
+
+    def remove_measuring_group_tab(self):
+        """Remove a tab in the Measuring group tab."""
+        self.dockwidget_controlled_structures.tab_measuring_group_view_2\
+            .removeTab(self.dockwidget_controlled_structures
+                       .tab_measuring_group_view_2.currentIndex())
+
+    def remove_all_measuring_group_tabs(self):
+        """Remove all tabs in the Measuring group tab."""
+        self.dockwidget_controlled_structures.tab_measuring_group_view_2\
+            .clear()